Switch to using locally defined bootstrap scss
Instead of using Bootstrap 3's system of only downloading what we need, we now use a custom file to dictate what bootstrap we need.

As a result, we don't have to re-download the css files if we decide we need another component, instead just add it to `custom.scss`.

Finally, this means we can use react-bootstrap components at will (on the condition that we add them to `custom.scss`